+++ app/build.gradle
@@ -18,8 +18,8 @@
 
 
 plugins {
-    id 'com.android.library'
-    //id 'com.android.application'
+    //id 'com.android.library'
+    id 'com.android.application'
 
 }
 
@@ -28,7 +28,7 @@ android {
     buildToolsVersion "30.0.3"
 
     defaultConfig {
-      //  applicationId "com.circo.oneflow"
+        applicationId "com.circo.oneflow"
         minSdkVersion 21
         targetSdkVersion 30
         versionCode 9
