+++ app/src/main/java/com/oneflow/analytics/OFFirstActivity.java
@@ -30,6 +30,7 @@ import android.os.Bundle;
 import android.view.View;
 import android.widget.EditText;
 
+import androidx.appcompat.app.AppCompatActivity;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
@@ -51,7 +52,7 @@ import java.util.Date;
 import java.util.HashMap;
 import java.util.List;
 
-public class OFFirstActivity extends OFSDKBaseActivity implements OFMyResponseHandler {
+public class OFFirstActivity extends AppCompatActivity implements OFMyResponseHandler {
 
     String tag = this.getClass().getName();
     OFCustomTextView result, sendLogsToAPI;
@@ -102,7 +103,7 @@ public class OFFirstActivity extends OFSDKBaseActivity implements OFMyResponseHa
 
         OFFontSetup optionsFont = new OFFontSetup(faceBold, 12f);*/
 
-        String projectKey = "oneflow_sandbox_oV+xY+hArzT2i4lMP69YZnRBLK1a/qmYW16MboVc208IVjiNKPfHRIylm0rVFgEubtaRuhKMTdlTt5TEuP+8Pw==";//"oneflow_prod_RyR/jsDNOiHS+GMW1ov0bykRA0NHE5mmIqM6eZJtN2ziWaecbiMQu+EvVDmmM3pUzupp7JJyZZcqZDlGASckiA==";
+        String projectKey = "oneflow_sandbox_7oKyqBl/myk8h1Zkq1uSk+8BmonEPLE7VLEhpgUD5FG8OpPFJ+nnLDVB5q6WSp60Os46wKO/gpmOmQwNwVgiaw==";//"oneflow_prod_RyR/jsDNOiHS+GMW1ov0bykRA0NHE5mmIqM6eZJtN2ziWaecbiMQu+EvVDmmM3pUzupp7JJyZZcqZDlGASckiA==";
 
         OneFlow.configure(getApplicationContext(), projectKey);//,titleSetup,descriptionFont,optionsFont);
         OneFlow.shouldShowSurvey(true);
