+++ app/src/main/java/com/oneflow/analytics/model/survey/OFSDKWidgetPosition.java
@@ -1,0 +1,39 @@
+package com.oneflow.analytics.model.survey;
+
+import com.google.gson.annotations.SerializedName;
+
+import java.io.Serializable;
+
+public class OFSDKWidgetPosition implements Serializable {
+
+    @SerializedName("is_enable")
+    Boolean is_enabled = true;
+    @SerializedName("mobile")
+    String mobile = "top-banner";
+    @SerializedName("desktop")
+    String desktop;
+
+    public Boolean getIs_enabled() {
+        return is_enabled;
+    }
+
+    public void setIs_enabled(Boolean is_enabled) {
+        this.is_enabled = is_enabled;
+    }
+
+    public String getMobile() {
+        return mobile;
+    }
+
+    public void setMobile(String mobile) {
+        this.mobile = mobile;
+    }
+
+    public String getDesktop() {
+        return desktop;
+    }
+
+    public void setDesktop(String desktop) {
+        this.desktop = desktop;
+    }
+}
