+++ app/src/main/res/anim/slide_enter_from_top.xml
@@ -19,5 +19,5 @@
 
 <translate xmlns:android="http://schemas.android.com/apk/res/android"
     android:duration="@android:integer/config_longAnimTime"
-    android:fromYDelta="50%p"
-    android:toYDelta="0%p" />
+    android:fromYDelta="-100%p"
+    android:toYDelta="10%p" />
