+++ app/src/main/res/anim/slide_exit_upward.xml
@@ -1,5 +1,4 @@
 <?xml version="1.0" encoding="utf-8"?>
-
 <!--
   ~ /* Copyright 2021 1Flow, Inc.
   ~
@@ -18,9 +17,8 @@
   ~
   -->
 
-<set xmlns:android="http://schemas.android.com/apk/res/android"
-android:shareInterpolator="true">
-<translate android:fromYDelta="0%" android:toYDelta="100%"
 
-    android:duration="600"/>
-</set>
+<translate xmlns:android="http://schemas.android.com/apk/res/android"
+    android:duration="@android:integer/config_longAnimTime"
+    android:fromYDelta="0%p"
+    android:toYDelta="-100%p" />
