+++ app/src/main/res/layout/survey_view_banner.xml
@@ -1,0 +1,107 @@
+<?xml version="1.0" encoding="utf-8"?><!--
+  ~ /* Copyright 2021 1Flow, Inc.
+  ~
+  ~
+  ~  Licensed under the Apache License, Version 2.0 (the "License");
+  ~  you may not use this file except in compliance with the License.
+  ~  You may obtain a copy of the License at
+  ~
+  ~       http://www.apache.org/licenses/LICENSE-2.0
+  ~
+  ~  Unless required by applicable law or agreed to in writing, software
+  ~  distributed under the License is distributed on an "AS IS" BASIS,
+  ~  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+  ~  See the License for the specific language governing permissions and
+  ~ limitations under the License.*/
+  ~
+  -->
+
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    android:id="@+id/base_popup_layout"
+    android:layout_width="match_parent"
+    android:layout_height="wrap_content">
+
+    <RelativeLayout
+        android:id="@+id/main_survey_child"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        
+        >
+
+    <RelativeLayout
+        android:id="@+id/slider_layout"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:padding="5dp"
+        android:visibility="gone">
+
+        <View
+            android:id="@+id/slider"
+            android:layout_width="50dp"
+            android:layout_height="5dp"
+            android:layout_centerHorizontal="true"
+            android:layout_marginTop="10dp"
+            android:background="@drawable/rounded_rectangle_small" />
+    </RelativeLayout>
+
+    <androidx.cardview.widget.CardView
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:layout_below="@id/slider_layout"
+        app:cardBackgroundColor="@color/plainwhite"
+        app:cardCornerRadius="0dp"
+        app:cardElevation="14dp"
+        app:cardUseCompatPadding="false">
+        <!--android:background="@drawable/main_rounded_rectangle_new_theme"-->
+        <RelativeLayout
+            android:id="@+id/view_layout"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:padding="10dp">
+
+
+            <RelativeLayout
+                android:id="@+id/inner_child_view"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content">
+
+
+                <!--<com.oneflow.tryskysdk.customwidgets.CustomViewPager
+                    android:id="@+id/survey_viewpager"
+                    android:layout_width="match_parent"
+                    android:layout_height="wrap_content"
+                    android:background="@android:color/white" />-->
+                <FrameLayout
+                    android:id="@+id/fragment_view"
+                    android:layout_width="match_parent"
+                    android:layout_height="wrap_content"
+                    />
+
+            </RelativeLayout>
+
+            <ImageView
+                android:id="@+id/close_btn_image_view"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:layout_alignParentRight="true"
+                android:layout_marginTop="2dp"
+                android:src="@drawable/ic_baseline_close_24" />
+
+        </RelativeLayout>
+
+        <ProgressBar
+            android:id="@+id/pbar"
+            style="?android:attr/progressBarStyleHorizontal"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:layout_below="@id/view_layout"
+            android:layout_gravity="bottom"
+            android:layout_marginBottom="-6dp"
+            android:indeterminate="false"
+            android:progress="100"
+            android:scaleY="1.2" />
+        <!--android:background="@color/plainwhite"-->
+    </androidx.cardview.widget.CardView>
+</RelativeLayout>
+</RelativeLayout>
